requirements.txt
torch==1.11.0+cu113 
torchvision==0.12.0+cu113
albumentations==1.0.3
wandb==0.10.33
matplotlib==3.4.1
seaborn==0.11.2
pillow==8.2.0
pandas==1.2.3
numpy==1.19.2
scikit-learn==1.0.2
scipy==1.6.2
hydra-core==1.1.0
fiftyone==0.14.3